refactor loadbalancerprofile - moves loadbalancer config manipulation code to pkg/utils/loadbalancer/ - moves loadbalancer test helper functions to pkg/utils/loadbalancer/ - updates references where appropriate
